Updated public chat group api service. move services to service folder. update program.cs with http client registration. Add struct to hold names for http clients.
