--- SignalRBlazorChatApp/HttpMethods/IPublicChatGroupsApiService.cs
@@ -1,13 +1,0 @@
-﻿using ChatApplicationModels.Dtos;
-using SignalRBlazorChatApp.Models;
-
-namespace SignalRBlazorChatApp.HttpMethods
-{
-	public interface IPublicChatGroupsApiService
-	{
-		Task<ApiResponse<PublicChatGroupsDto>> DeleteGroup(int groupId, string jsonWebToken);
-		Task<ApiResponse<List<PublicChatGroupsDto>>> GetPublicChatGroupsAsync(string jsonWebToken);
-		Task<ApiResponse<PublicChatGroupsDto>> PostNewGroup(CreatePublicChatGroupDto createDto, string jsonWebToken);
-		Task<ApiResponse<PublicChatGroupsDto>> UpdateGroup(ModifyPublicChatGroupDto modifyDto, string jsonWebToken);
-	}
-}
