--- SignalRBlazorChatApp/HttpMethods/PrivateGroupMessagesApiService.cs
@@ -1,121 +1,0 @@
-﻿using ChatApplicationModels.Dtos;
-using Microsoft.AspNetCore.WebUtilities;
-using Newtonsoft.Json;
-using SignalRBlazorChatApp.Models;
-using System.Net.Http.Headers;
-using System.Text;
-
-namespace SignalRBlazorChatApp.HttpMethods
-{
-	public class PrivateGroupMessagesApiService : IPrivateGroupMessagesApiService
-	{
-		private HttpClient _httpClient;
-		private readonly IConfiguration _configuration;
-
-		public PrivateGroupMessagesApiService(HttpClient httpClient, IConfiguration configuration)
-		{
-			_httpClient = httpClient ?? throw new Exception(nameof(httpClient));
-			_configuration = configuration ?? throw new Exception(nameof(configuration));
-		}
-
-		public async Task<ApiResponse<List<PrivateGroupMessageDto>>> GetMessagesByGroupId(
-			int groupId, int numberMessagesToSkip, string jsonWebToken)
-		{
-			_httpClient = GetNewHttpClient(_httpClient, jsonWebToken);
-
-			var query = new Dictionary<string, string>()
-			{
-				["groupId"] = groupId.ToString(),
-				["numberMessagesToSkip"] = numberMessagesToSkip.ToString()
-			};
-
-			string baseUri = BaseUri();
-			var uriWithQuery = QueryHelpers.AddQueryString($"{baseUri}bygroupid", query!);
-			var dataRequest = await _httpClient.GetAsync(uriWithQuery);
-
-			string jsonContent = await dataRequest.Content.ReadAsStringAsync();
-			ApiResponse<List<PrivateGroupMessageDto>> apiResponse = JsonConvert
-				.DeserializeObject<ApiResponse<List<PrivateGroupMessageDto>>>(jsonContent)!;
-
-			return apiResponse;
-		}
-
-		public async Task<ApiResponse<PrivateGroupMessageDto>> PostNewMessage(
-			CreatePrivateGroupMessageDto createDto, string jsonWebToken)
-		{
-			_httpClient = GetNewHttpClient(_httpClient, jsonWebToken);
-
-			string baseUri = BaseUri();
-
-			var bodyMessage = new StringContent(
-				JsonConvert.SerializeObject(createDto), Encoding.UTF8, "application/json");
-
-			var postRequest = await _httpClient.PostAsync(baseUri, bodyMessage);
-
-			string jsonContent = await postRequest.Content.ReadAsStringAsync();
-			ApiResponse<PrivateGroupMessageDto> apiResponse = JsonConvert
-				.DeserializeObject<ApiResponse<PrivateGroupMessageDto>>(jsonContent)!;
-
-			return apiResponse;
-		}
-
-		public async Task<ApiResponse<PrivateGroupMessageDto>> UpdateMessage(
-			ModifyPrivateGroupMessageDto modifyDto, string jsonWebToken)
-		{
-			_httpClient = GetNewHttpClient(_httpClient, jsonWebToken);
-
-			string baseUri = BaseUri();
-
-			var bodyMessage = new StringContent(
-				JsonConvert.SerializeObject(modifyDto), Encoding.UTF8, "application/json");
-
-			var updateRequest = await _httpClient.PutAsync(baseUri, bodyMessage);
-
-			string jsonContent = await updateRequest.Content.ReadAsStringAsync();
-			ApiResponse<PrivateGroupMessageDto> apiResponse = JsonConvert
-				.DeserializeObject<ApiResponse<PrivateGroupMessageDto>>(jsonContent)!;
-
-			return apiResponse;
-		}
-
-		public async Task<ApiResponse<PrivateGroupMessageDto>> DeleteMessage(Guid messageId, string jsonWebToken)
-		{
-			_httpClient = GetNewHttpClient(_httpClient, jsonWebToken);
-
-			var query = new Dictionary<string, string>()
-			{
-				["privateMessageId"] = messageId.ToString()
-			};
-
-			string baseUri = BaseUri();
-			var uriWithQuery = QueryHelpers.AddQueryString($"{baseUri}", query!);
-			var deleteRequest = await _httpClient.DeleteAsync(uriWithQuery);
-
-			if (!deleteRequest.IsSuccessStatusCode)
-			{
-				string jsonContent = await deleteRequest.Content.ReadAsStringAsync();
-				ApiResponse<PrivateGroupMessageDto> apiResponse = JsonConvert
-					.DeserializeObject<ApiResponse<PrivateGroupMessageDto>>(jsonContent)!;
-				return apiResponse;
-			}
-
-			return new() { Success = true, Message = "ok" };
-		}
-
-		#region PRIVATE METHODS
-
-		private string BaseUri()
-		{
-			return _configuration["ApiEndpointsConfig:PrivateGroupMessagesUri"]!;
-		}
-
-		private HttpClient GetNewHttpClient(HttpClient httpClient, string jsonWebToken)
-		{
-			httpClient = HttpClientFactory.Create();
-			httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", jsonWebToken);
-			return httpClient;
-		}
-
-		#endregion
-	}
-}
