+++ SignalRBlazorChatApp/SignalRBlazorChatApp.csproj
@@ -43,7 +43,6 @@
 
   <ItemGroup>
     <Folder Include="Logs\" />
-    <Folder Include="Services\" />
   </ItemGroup>
 
 </Project>
